[clock][core] Don't treat QuerySupportedRate::NotFound as an error.
Sometimes a client will query a clock device for a supported rate and the
clock device will not be able to satisfy the request. This is not an
error condition and we reserve the `ZX_ERR_OUT_OF_RANGE` status code to
signal this. In the future we may opt for a more semantic mechanism to
signal this. See b/426652785.

Bug: b/425542452
Bug: b/425942556
Change-Id: If3b810c01efdec81ad062bfbd9633905f10e92b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1308068
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>