[kernel][lib/arch][arm] Use __arm_mrrc instead of __arm_rsr64
The __arm_mrrc intrinsic maps directly to the armv7 instruction
that the aarch64-compatible __arm_rsr64 intrinsic also maps to on
armv7.  But the compiler seems to treat it more faithfully as a
volatile read, and it's a more direct way to express the intent.

Change-Id: I7096993cd397e460190ac1434d0fb71bac97d2d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1305884
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>