[fshost] Stop opening the unencrypted volume twice
It was possible for the unmount for the first open to be slow and race
with the second open and cause failures.

Also, explicitly shut down fxfs when shredding. I think this isn't
strictly necessary because we explicitly unmount the unencrypted
volume. However, I think this is a bit safer.

Fixed: 427473055
Bug: 427478050
Change-Id: I8f8fbd3784b2f51582006ac5f7ba3edf46c18795
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1308487
Reviewed-by: Rui Qi Sim <rqsim@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>