[suspend][RFC] Create a RFC for Wake Source Reporting
Create an RFC to propose a method for "wake source reporting"; a way
to provide specific details to users of `zx_system_suspend_enter` used
to understand all of the specific reasons why the system eventually
exited the suspended state and returned control to user mode.

This information in meant to facilitate debugging and diagnostic
scenarios, as well provide the ability of the Fuchsia power management
subsystem to satisfy the reporting requirements of systems built on
top of Fuchsia.

Bugs: 374824168, 374827864
Change-Id: I0d84c1a4d36a56beb65a216e53a6d3e2b9552a31
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283694
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Reviewed-by: James Robinson <jamesr@google.com>