[starnix] Mark a hrtimer that is associated with an interval timer
Ensures tha a hrtimer that backs a Starnix interval timer is correctly
marked as an interval hrtimer. Interval hrtimers are treated specially,
as they prevent Starnix container suspension if they are in a "pending
reschedule" state.

By all accounts the hrtimers used in interval_timer.rs *should* be
marked as interval hrtimers, but they weren't.

I reshuffled some code to ensure that when an interval timer is
scheduled, it is properly marked. The reshuffle does not seem to
change the semantics of the other effects of the code, even though
the initialization ordering is shuffled around a bit. An important
part is the change of the `is_interval` value on a hrtimer, which
affects its wake alarm identifier, and should not be modified while
the hrtimer is running. Since `arm()` call implies hrtimer is not
running, we should be OK to make the change here.

A possibly more robust approach would be to create the hrtimer as
interval, but that's a more significant change that I'd like to avoid in
this first instance.

Multiply: sysfs_power_test
Bug: 425669944
Change-Id: I4c3e71974e4b9f18086ff4f135125c7895d5b74e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1309186
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>