[kernel] Move stack guard selection into lk_main()
Eventually this will happen properly in physboot when it has secure
access to a source of randomness (e.g., for KASLR support). In the
meantime though this doesn't have to happen in start.S. ABI set-up can
initialize the thread pointer with a bogus, temporary guard, and then
can reinitialize it properly in C++ soon after entering lk_main().

This change is part of the pruning of start.S and simplifies the
migration of ABI set-up that's to happen in physboot.

Bug: 42164859
Change-Id: I892468c3c6383b49372b64e082a94f4b99d360aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212632
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>