[kernel][phys] Introduce phys handoff requirements
This information about the kernel stage needed at the phys stage to
properly prepare it. The expectation is that structure will be recorded
as an ELF note in the kernel to be loaded and parsed during hand-off
preparation. This will come to contain the things relating to ABI set-up
(like stack sizes), but for now is an empty struct with this change
validating simply that it is present.

Bug: 42164859
Change-Id: I02ce11a2d073480191843542f1e3b0a1d6716dfa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1306905
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>