[kernel][zbi] Introduce zbi_dcfg_riscv_plic_driver_t::size_bytes
It's generally good practice for kernel driver configuration to give the
full extent of its MMIO range and not just the base address, especially
if statically described in devicetrees. We update the PLIC config to do
this for the additional reason that the PLIC driver is the final
(officially supported) kernel driver that accesses its MMIO through the
physmap (as it comes up before the VM subsystem), and the inclusion of
length in the config allows physboot to instead prepare and hand-off a
dedicated mapping.

Bug: 42124648
Change-Id: Id7d1a42775f85d0289d71f70db0be7ce69252669
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1309484
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>