[zx][rust] Implements a mappable clock wrapper
This is a follow on to https://fxrev.dev/1215126, except
it takes a few more design decisions to eliminate unsafe
method use, and hopefully make for a more ergonomic API.

Moved the methods outside of zx, since this is not really a
separate kernel object.

To be used in Starnix.

Bug: 422533641
Change-Id: I0f8e93e5ec40a8e2cc03e81b44896ddbc2c1dd1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295094
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>