[starnix] Restructure task priority/niceness
No behavioral change; this supports upcoming https://fxrev.dev/1286324
which will fix a smörgåsbord of defects.

We don't get to have the data layout that we might like to have (with
real-time priority only being present during use of a real-time
policy); in SchedulerState by placing realtime_priority as a sibling of
policy we violate the principle of "don't store state that is entirely
determinable from other already-stored state (unless you are caching)"
(this principle is not the same as, but lives next door to, the more
famous "make invalid state unrepresentable"). For the moment this shape
improves the surrounding code; perhaps later on we'll find a way to
right the data representation without degrading the surrounding code.

Bug: 397371605
Change-Id: I1c8f12b0232968c0d8c22a173ca6895d4a923c2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1305364
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Fuchsia-Auto-Submit: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>