[starnix] fix access checks for /dev/kmsg
This change adds tests for /dev/kmsg's behavior, and fixes the
mismatches between Starnix and Linux:
 - opening /dev/kmsg for writing only requires a regular file access
   check
 - opening for reading does not involve a capability check if
   dmesg_restrict is false
 - CAP_SYS_ADMIN cannot replace CAP_SYSLOG in this case
 - a SELinux access check for system { syslog_read } is required
The tests verify that permission checks are only applied on opening.
We remove the SyslogNonRootTest.DevKmsg test: all tests scenarios are
covered by the new tests, and the existing test was too dependent on
the mode for the device in /dev/kmsg (which ends up being 0666 with
our current devtmpfs, but is 0644 on a Linux workstation and 0600 on
Android).

Bug: 403532599
Change-Id: I220f56628a2dcfe3a752f9b089261006f42700aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304324
Commit-Queue: Ambre Williams <ambre@google.com>
Reviewed-by: Wez <wez@google.com>