[fshost] Remove Admin.GetDevicePath
This method is not implemented for our new fshost implementation, and
there is a desire to avoid using topological paths. The only use case
was for the Fuchsia `df` shell utility. For now, we remove the "Device"
column from `df` output, since it was always set to "Unknown". If there
is a need to provide a path to the device in the future, we can consider
having the `fuchsia.io/Node.QueryFilesystem` method return the
topological path instead.

Remove an unrelated GetDevicePath method from GPT tests that was unused.

Bug: 377560660
Test: Built + ran new `df` utility.
Change-Id: Ib8bd6ef026330a4237bba05fd8f52f49cb32ed6b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1307645
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>