[sestarnix] listen for avc audit messages in Linux userspace tests
The printing of audit messages to klog is greatly limited on Linux,
since this is supposed to be a fallback mechanism until an audit
daemon comes up. To get the full avc output, we spawn a process that
listens for avc messages on netlink audit and prints them to the
console.

Change-Id: I4e048cc1ebe1df4cea919301e7eef510c3b25777
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1306464
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Nathaniel Manista <nathaniel@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>