[fshost] Create separate recovery vs. non-recovery shards
Split fshost protocol routing such that the Recovery protocol is
optional, and can be configured by assembly. This allows us to build
products that don't allow accessing recovery functionality.

This protocol is also used by the paver service, which is currently
included in every build type that's based on bootstrap or higher. In the
future however, it's likely that we would want to assemble a product
image without either recovery or paving capabilities.

Bug: 377560660
Change-Id: I0acf766b8ce916ef150a0df4fa8d1a29e89ff7b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1309284
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>