[perf] Map frame event pids to Fuchsia koid
This parses the protobufs read from Perfetto and maps the pid
in the data to the corresponding Fuchsia koid, which is used as the
the process id Fuchsia.

When the trace is stopping, just before the perfetto data is read,
the pid table is copied from the kernel to record pid and
koid informnation for all tasks. This assumes that all processes that
emit linux events are still running when the trace stops.

A key assumption is also that pids will not be reused for the duration
of the trace.

The result is perfetto track data that is identified using the pid will
be associated with the track data generated by Fuchsia using the KOID as
the process id.

Bug: 319492968
Change-Id: I15a9ae5c7e133b3a07e629b816f1f67219fecfa8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1281445
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>