[starnix][sysfs] Normalize kernel and power directories
Previously, the "build" functions for these directories created the
directories themselves. The pattern that has developed for sysfs is that
the "build" function should build the inside of the directory and the
caller should be responsible for creating the directory itself. This CL
updates these older build functions to match the modern pattern.

Bug: b/419306849
Change-Id: I0a0d5f44552baf446e3207d637946dfde2bc631d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1310855
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>