[zircon][scheduler] Add flag for unified bookkeeping
This CL introduces a build-time for unified bookkeeping in the Zircon
scheduler. This flag will let us land the code for unified bookkeeping
in a disabled state so that we can enable it on bots and work through
the race conditions that this new scheduler reveals in other code.

Bug: b/428694298
Change-Id: I3b251a558f06d7933c1601aa4e75127cdfe93cc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1311984
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>