[starnix][hrtimer] Rename `init_for_test` to `init_internal`
The "for test" naming was supposed to indicate that this method should
not be called directly in prod code.

But this is ambiguous, since at a cursory glance may look like a
testonly code, which it is not.

Bug: 428223204
Change-Id: Ie073b0d7afe808a2b4a18a0661f2961e62cd6feb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1311224
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>