[fuchsia-controller][docs] Update `Context` usage. Remove daemon references.
The daemon is no longer used by `fuchsia-controller` and hasn't for some
time. This is reflected in the doc. Furthermore I've added a section
outlining the use of the `Context` object as it is central for
`fuchsia-controller`'s functionality.

This also updates some existing documentation for the `ffx` runtime
configuration by showing the usage of the dot notation.


Change-Id: I8906bb31a80824003328d8d80babed296a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312123
Reviewed-by: Kyo Lee <kyol@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>