Reland "[bazel_sdk][tests] Migrate deps to bzlmod"
This is a reland of commit 6d8fc17e0188ac73888bafb4b7ef263b90dfe0f1

More context for future references:

This was one of the changes in a stack that was reverted because
https://fxrev.dev/1300384 broke fuchsia_sdk roll OOT. The change
causing the breakage has been successfully relaned in
https://fxrev.dev/1309764 and tested in http://ag/34249619.

I'm also seeing a roll of fuchsia_sdk that picks up the change, which
passes presubmits in  http://ag/34287100.

Concluding it should be safe to continue relanding the rest of the chain
that was reverted.

Original change's description:
> [bazel_sdk][tests] Migrate deps to bzlmod
>
> Bug: 418319048
>
> Change-Id: I46f009e19b17710a6d07e3be027065263bdc39e3
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1304825
> Reviewed-by: David Turner <digit@google.com>
> Commit-Queue: Jay Zhuang <jayzhuang@google.com>

Bug: 418319048
Change-Id: If33c26d6b7ca754fdd32bcf333e857906af2e298
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1311664
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Jay Zhuang <jayzhuang@google.com>