[test-pilot] move output definitions to lib
This CL moves the definitions for two output types (the summary and the
invocation log) to a separate library, allowing other tools (e.g. ffx
test) to use test-pilot's definitions. Note that this has not been done
for test configuration, because different tools will be reference
different parameters in the configuration, and because the configuration
is just a flat list of parameters.

This CL also makes some updates to error formats, so that error sources
are displayed.

Test: no significant behavior change

Bug: 327640651
Change-Id: Icbb18028d35a7de749a572c7fba8c845ba62ec08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312468
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>