[netstack2] Reduce log level of FIDL error
This FIDL error is caused by a race in the golang FIDL bindings. By
logging at `Error` it's causing test flakes. Reduce this to `Warn` to
prevent test flakes.

Netstack2 is in maintenance mode, and it's the only consumer of the
golang FIDL bindings, so it's not worth identifying + fixing the
underlying race.

Fixed: 428338722
Change-Id: Ia5f8d07aed0436779e059cbc37ddeb8af2822d2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1311924
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>