[kernel][phys][riscv64] Handoff PLIC MMIO mapping
The PLIC driver accesses its MMIO through the physmap. It did this as an
expedient measure since it came up before the VM subsystem, we didn't
yet have a periphmap-like mechanism on RISC-V, and physboot was not
ready yet. Now that physboot is ready (and that the PLIC config includes
MMIO range length) we have it prepare and hand-off a dedicated MMIO
mapping.

This was the last officially supported kernel driver to access MMIO
through the physmap.

Bug: 42124648
Change-Id: I6d41eb7f355ca67d5bcbfeebeff643483ff1163b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1309485
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>