[fidl][rust] Replace extension traits with deref
One of the goals for the new Rust bindings is to avoid the
protocol-specific newtypes, since these tend to make writing generic
code more difficult. However, this conflicts with Rust's orphan rules
which prevent us from writing inherent impls for types defined in other
crates. Extension traits were a decent solution to this problem, but
they have to be in scope for their methods to be available.

An alternative solution to this problem is using Rust's deref coercion.
By creating a newtype and writing a `Deref` impl for the generic
`*Sender` types, we can provide the protocol-specific methods without
requiring users to bring additional items in scope. This makes it easier
to work ergonomically with the sender types.

Change-Id: Ibba93b36b0c05468aef4365006332b44bced3bab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1303486
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: David Koloski <dkoloski@google.com>