[netstack3] Add Counter for RxQueueFull on sockets
Increment a counter whenever we fail to deliver a message to a socket
because it's receive queue is full.

Even though the receive queue is in bindings, I opted to put the counter
in core: core already has all the infrastructure in place to support
per-socket counters.

Bug: 413418876
Change-Id: I252c2254e383f70ca15241cc0cbbd807ecc565e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1311144
Commit-Queue: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Fuchsia-Auto-Submit: Jeff Martin <martinjeffrey@google.com>