[starnix] Refactor locked parameters in test
Instead of getting an instance, tests now directly get a mutable
reference. This allows not to have to convert it to a mutable reference
at the point of use, and so follow the same pattern as the non test
code.

Change-Id: I57aba170daf9389f9341e9de3c6cabe238b374a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312764
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>