[kernel][phys] Introduce HandoffPrep::ArchDoHandoff()
We introduce the final, arch-specific routine for branching into the
kernel at the end of hand-off. This was previously accomplished as a
plain function call to _start, but this was only ever intended as a
placeholder; we'll soon need sites to final arch-specific preparation
(e.g., stack installation) in assembly.

We make this routine subsume HandoffPrep::ArchHandoff(), which currently
just makes a few ArchPhysHandoff settings that can happen at any point;
we update things to just do that logic at the top of ArchDoHandoff().

Bug: 42164859
Change-Id: Idaf8cf4880eaae87e3e712d56a319a43fbe50a7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1310095
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>