[sdk] Add `transitive_features` attribute to fuchsia_cc_binary() and fuchsia_cc_test().
This is a reland of [1] which got reverted because the HSP out-of-tree
build still uses Bazel 6.4 which requires an allowlist label when using
transitions (see https://bazel.build/versions/7.4.0/extending/config).

This patch contains the fix, compared to patchset 1 to see what changed
from the original CL.

== Original commit message ==

Add a new attribute to the fuchsia_cc_xxx() macros to provide a list of
feature flag names to apply to the C++ binary and all its transitive
dependencies. This injects a rule to apply a transition modifying the build
configuration's features list (i.e. `//command_line_option:features`).

This uses a custom rule that can export the CcInfo and CcSharedLibraryInfo
providers from the real C++ binary target, as they are needed by the
fuchsia_cc() rule to attach extra

This is necessary to disable sanitizer features when building certain
targets. The root problem is that due to b/417952817, sanitizer flags are
not passed properly to Bazel when invoked from GN, and trying to fix this
reveals that some //vendor/* Bazel tests are failing at runtime with
undefined behavior issues.

See tqr/1005170 which uses this new attribute to disable asan and ubsan
for the problematic target, then https://fxrev.dev/1269985 which fixes the
flag-passing bug.

Tested also with `fx bazel build @prebuilt_clang//:test_suite`.

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1301585

Bug: 417952817
Bug: 425323223

Change-Id: I459ab4b0395776386dba54158277d2c259e0d85b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1309964
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Turner <digit@google.com>