[starnix] Fix deadlock in ptrace
ptrace could deadlock because it was trying taking a lock on the pid table
while having a TempRef of task. Because the task would take a lock on
the pid table before waiting that all temp ref are release, this
introduced a race.

Bug: b/420334742
Change-Id: I11da9712e9b41bd291bee4fdbd813bf5facc4776
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1313564
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>