[bt][rfcomm] Reorder RX datapath
Previously, when an incoming  packet was received, bt-rfcomm would
update the local credit count, attempt to send any queued TX packets
that were blocked on credits, and then notify the application of the
received RX packet. This prioritizes the TX path at the tradeoff of
latency for the application layer.

Update to immediately relay the packet to the application. This is a
non-blocking operation so it should not affect the TX latency by
much even if the application is slow.

Bug: 371346522
Test: fx test bt-rfcomm-tests
Test: Manual, OOBE

Change-Id: I081f70e7a1b116a8641ff5683096603a73bf800d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1310426
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>