[roll] Roll llvm-project/libc [libc] Fix internal alignment in allcoator (#146738)
Summary:
The allocator interface is supposed to have 16 byte alignment (to keep
it consistent with the CPU allocator. We could probably drop this to 8
if desires.) But this was not enforced because the number of bytes used
for the bitfield sometimes resulted in alignment of 8 instead of 16.
Explicitly align the number of bytes to be a multiple of 16 even if
unused.
GitOrigin-RevId: 50f40a5327ad9f8a4a57bb2bf7679f489f86b726
Original-Revision: d728755b5d048f6773bf2bfed548d0eaad094d52

Rolled-Repo: https://llvm.googlesource.com/llvm-project/libc
Rolled-Commits: c076cd7c4624f0..d728755b5d048f
Roller-URL: https://cr-buildbucket.appspot.com/build/8710436076650004001
CQ-Do-Not-Cancel-Tryjobs: true
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I28325d373c677806b37c3aecb5e0a1101dd17ca6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1313924
Commit-Queue: GI Roller <global-integration-roller@fuchsia-infra.iam.gserviceaccount.com>