[sestarnix] Apply the ForClass trait
Common permissions for file system nodes, files and sockets are
now passed to the has_*_permission() helpers without first
requiring specialization to a class-affine KernelPermission.

This removes the need for the FsNodeLabel to be locked/unlocked
separately, to determine the class and effective SID, as well as
simplifying the call-sites slightly.

Bug: 423838496
Change-Id: I8c6c3b5b8ff316a77d4c97b6a12f3ff71bd03103
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1306645
Commit-Queue: Wez <wez@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Marina Ciocea <marinaciocea@google.com>