[bt][rfcomm] Adjust credit watermark levels
The current max watermark (255) is quite large when considering a
resource-limited system. This can result in a maximum of
255*max_packet_size bytes received from the remote peer. In practice,
the max_packet_size tends to be around ~1000 bytes, so this would max
out at ~255KB.

Adjust the maximum watermark for remote credits to be 16. This is
sufficient for high-throughput operations (e.g. OOBE) while providing
ample headroom for a resource limited system.

Adjust the minimum credit watermark to be 4 to reduce the frequency of
sending an empty data frame for the sake of a credit refresh. Credit
refreshes will now typically occur during the standard TX datapath,
instead of out-of-band.

Bug: 428738649
Test: fx test bt-rfcomm-tests
Test: Manual OOBE

Change-Id: I984582eb4ca0fda50fba9d4605c95cccd4403c90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1309362
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>