[trace] Display ftrace/print events with correct process
This map the process id of ftrace/print events in the
Perfetto trace data stream to the underlying Koid of the
process if present.

This makes the events appear in the correct process.

Long term, this code will not be needed since we want to
make Perfetto traces contain all Fuchsia trace data.

Example trace:
https://ui.perfetto.dev/#!/?s=36784295f1627a6adabc2788b9df7dbb98216e9c

Bug: 402192254
Change-Id: Id52bc5f691785765a1d60ab4a39fef56942e840d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312385
Reviewed-by: Jia Li <lizhijia@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>