[starnix] Move did_create_dir_entry hook to factory method
This change moves the `did_create_dir_entry` hook to the
`DirEntry::new` factory method.

By placing the hook in the central factory method, we guarantee that
it is executed for every `DirEntry` that is instantiated,
regardless of the code path that leads to its creation.

Previously, the hook was called in a more specific creation path,
which could be bypassed, leading to inconsistent behavior. This
change makes the hook more robust and ensures that filesystem-specific
logic is always applied upon entry creation.

Change-Id: Ic15798695bf6cab1973017a1da4f7667e34f876d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1315804
Reviewed-by: Maria Glukhova <mariagl@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>