[starnix] Add network_manager to config
Add a feature flag to enable transitioning to the enablement of
socket_proxy and socket marking. This must be enabled alongside the
socket mark and the socket proxy must be included in the build.

This is similar to the ifb0 config change in fxr/1283704.

Bug: 428745577
Change-Id: I43e6116d2a3b9a005ddf85afa2b2cabefa31a5aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312118
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>