[fuchsia-async] Inline OnSignals allocation
This commit includes four major changes:

1. Changing the lock around fasync's `PacketReceiverMap` to an `RWLock`
2. Switching `PacketReceiverMap` to hold data pointers and a vtable.
3. Introducing `RawReceiverRegistration` to allow for more efficient
   packet receiver registration
4. Splitting the core implementation of `OnSignals` into an unboxed
   `OnSignalsFuture` and a boxed `OnSignals`.

`OnSignals` only ever registers a packet receiver while being polled,
and therefore only when it is pinned. Since it's pinned, we can avoid
making an additional allocation for the coordinating memory between the
executor and the future. Being pinned guarantees that `drop` will be
called to unregister the receiver before the memory is freed.

To prevent a race removing raw packet receivers, raw packet receivers
are now called while holding a read lock on the packet receiver
map. This does open some potential for deadlocks if packet receivers
attempt to insert or remove packet receivers during their
callback. Owned packet receivers, which incur a heap allocation, do not
have this limitation.

As a side-effect of some code reorganization, this change also switches
a few `Arc`s to `Box`es. This should be vanishingly more efficient.

In benchmarks, this removed an allocation corresponding with each
`OnSignals` created. It also reduced the amount of memory allocated
slightly, by about 1% overall. This is not expected to be representative
of the entire codebase.

Co-authored-by: Chris Suter <csuter@google.com>
Change-Id: I3863688cd91508bd07140fe727c9a8abf6b5a0db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1313267
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>