[botanist] Use -serial-log-dir for emulators instead of logfile.
The recipes specify the serial log path in the emulator config to tell
botanist to just copy the emulator output to the serial log instead of
reading from the serial socket which won't be available unless the
`serial` field is set to true in the emulator config. It also makes sure
the -serial-log-dir flag is not set if `serial`=false. Instead we should
just use that `serial` field to tell us which way to capture the serial
log and unconditionally set the -serial-log-dir flag instead of using
the logfile in the config.

Bug: 422604948
Change-Id: I74ef111bd8cba0c0d799c30c16c8824a36beb3f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1314271
Reviewed-by: Carver Forbes <carverforbes@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>