[starnix] Add `re_own` to `WeakRef` and `TempRef`
This change introduces a `re_own` method that allows upgrading a `WeakRef`
or `TempRef` back to an `OwnedRef`.

The `re_own` method will only succeed if there is at least one existing
`OwnedRef` to the object, ensuring that ownership is not incorrectly
resurrected.

This change also includes unit and concurrency tests to verify the
correctness and thread-safety of the new functionality.

Change-Id: I5ea8dd0fe0062e49ffdc05a22dfedd6525c0da1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317124
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>