[starnix][vfs] Rename DirEntryCache
Previously, it was not obvious how the `node_cache` differed from the
`entries` in the FileSystem struct. This change makes it clear from the
types that the `node_cache` is an `FsNodeCache` and the `dcache` is a
`DirEntryCache`. Also, this CL renames `entries` to `dcache` to match
the common name for this cache in other VFS implementations.

Change-Id: I16bc0d9a4ddfb89088abd91eeabe9a3dc2c24f09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317407
Commit-Queue: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>