[build][idk] Require hdrs and api_area in idk_cc_source_library()
It doesn't make sense to have a source library without headers, and
stable libraries already require headers to populate `api_contents_map`.
This also avoids putting all files in `srcs` as was often the case with
the GN `sdk_source_set()` template.

Require `api_area` so that it gets populated as libraries are migrated.
Add support for it to bazel2gn, and specify the area in existing uses
of this macro.

Also add "Required" comments to the descriptions of legacy macro
arguments.

Bug: 417305295
Change-Id: I7d6eae1479a7e71cf5f97b86700e0fbdce85bb5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312384
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>