[netpol] Transmit Fuchsia networks to SocketProxy
Connect to the FuchsiaNetworks protocol inside netcfg to share
updates from Fuchsia-provisioned networks.

When a network comes online with a v4 or v6 default route, add it to the
FuchsiaNetworks known by the socketproxy. When a network loses all of
its default routes, remove it from the networks known by the
socketproxy.

Updates to the netstack-dns-integration-tests will be uploaded in a
separate CL. They continue to work unmodified at the moment because
the tests interact directly with FuchsiaNetworks and not with
netcfg-installed interfaces.

Bug: 388113712
Fixed: 385368910

Change-Id: I4450ecace971d5963c328858d92778eb8fb093c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1165212
Reviewed-by: Zoey Bush <zmbush@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>