[compat] Replace fragment validation in fidl connect path
We enumerate parent fragments by listing incoming instances of
fuchsia.driver.compat.Service. The parent should not be required to
implement this service if it doesn't serve banjo or metadata to the
child so we shouldn't rely on it to validate the connection logic.

The fdf::ServiceValidator does the correct thing here so we use that
instead.

Change-Id: I21fe7349a89831af2c768c3759e2929704a70414
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1317550
Reviewed-by: Novin Changizi <novinc@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Suraj Malhotra <surajmalhotra@google.com>