[i2c-channel] Rename library
Rename the //src/devices/i2c/lib/device-protocol-i2c-channel library to
"i2c-channel-legacy". This is to denote that this library is the legacy
version and should not be used. A future CL will create a non-legacy
version of the library compatible with DFv2 drivers.

Bug: 430034481
Change-Id: Ibe9e5704b8ddc094185d154c986b9a12a2bdee12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1316104
Fuchsia-Auto-Submit: Nick Eaton <nickeaton@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>