[starnix] Add a `spawn_async` for kernel threads
This CL adds a new `spawn_async` method to `KernelThreads` that allows
spawning a kernel thread from an `async` block.

This is implemented by having the `spawn` method run a `LocalExecutor` on
the new thread, which in turn runs the `async` block.

A new `LockedAndTask` struct is also added to allow passing both a
`Locked` and a `CurrentTask` to the `async` block.

The new `spawn_async` method is used in a few places to simplify the
code.

Change-Id: Ib5f135eab24b8c7098d1850455a6eb9e061bb791
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1318264
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>