[testsharder] Fix calculation of target cpu.
Before we were using the cpu from the test spec of the first test in the
shard. However, for host tests, the cpu is the cpu of the host, not the
target. This is only a problem where the target cpu doesn't match the
host cpu such as for arm64 with tcg and only for emulators which make
use of this value. So far we'd only been using the TargetCPU() to
determine the HostCPU() which ended up not having a problem, but with
fxrev.dev/1316305 we started using it in the botanist config so it was
giving the wrong value.

Bug: 422604948
Change-Id: Ifdbab6243b0ed3d8be19f7d5260b0f719a4f5385
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1318885
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Carver Forbes <carverforbes@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>