[trace][ffx] Refactor target_info/config out of TraceTask
This moves the references to target_info outside the TraceTask struct
and keeps them in the daemon code.

It also moves read the ffx_config data to expand category groups
outside the TraceTask.

This in preparation for reusing the TraceTask on device to implement
detached tracing.

Bug: 428951658
Change-Id: Ib3ec78c994a8853de648fbd1bd11dbc32a1caab5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1312926
Reviewed-by: Andrew Davies <awdavies@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>