[unwinder] Support ARM EH Unwinding ABI
Adds a new unwinder, module, and parser classes that mimic the CFI
unwinder implementation for the Arm Exception Handling ABI.

The module is responsible for loading and performing a binary search
through the ARM.exidx binary search table, which is then passed off to
the parser which is responsible for decoding unwinding instructions.

This is not yet a complete implementation, with lots of opcodes still
not being handled as well as non-inlined unwinding instructions in the
ARM.extab section, but this is sufficient to get reasonable looking
backtraces for 32 bit binaries.

Tests are difficult to write at the moment due to the existing testing
strategy not allowing us to collect backtraces of 32 bit binaries. A
follow up will add limited testing coverage to zxdb_e2e_tests, and full
testing support will be covered in https://fxbug.dev/430036274.

Bug: 395606674
Bug: 416040479
Bug: 430036274
Change-Id: Ia7d1814064e14bb922a0eb7b59cd010c5379950c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272195
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>